Fix rebase-introduced import cycle and stale config test
Two regressions surfaced after rebasing the dual-name env-var change onto
main:

- containers.py imported add_bottle_env_aliases from data.py at module
  scope, but data.py already imports gateway constants from containers.py,
  closing a circular import (data -> containers is the established
  direction).  That broke collection of the entire compute_space test
  suite.  Move the alias helper and its prefix constants into a new
  dependency-free module, core/env_aliases.py; containers imports from
  there, and data re-exports the names (it remains the historical home of
  the OpenHost -> Cloud in a Bottle mirror boundary).

- test_bottle_env_aliases.py asserted on Config.zone_domain, which main
  removed ("move domain out of config again" — zone_domain now lives in the
  DB, not the config).  Repoint the affected test at
  my_openhost_redirect_domain, a config field that still exists, and drop
  the now-meaningless OPENHOST_ZONE_DOMAIN setup from its sibling.
