+++ compute_space/src/compute_space/core/data.py
@@ -8,30 +8,17 @@ import subprocess
 
 from compute_space.core.containers import ROUTER_GATEWAY_HOST
 from compute_space.core.containers import ROUTER_LOOPBACK_HOST
+
+# The env-var alias helpers live in ``core.env_aliases`` (a dependency-free
+# module) to avoid an import cycle with ``core.containers``.  They are
+# re-exported here (redundant ``as`` alias marks the intentional re-export)
+# because ``data`` is the historical home of the OpenHost -> Cloud in a Bottle
+# mirror boundary.
+from compute_space.core.env_aliases import ENV_PREFIX as ENV_PREFIX
+from compute_space.core.env_aliases import LEGACY_ENV_PREFIX as LEGACY_ENV_PREFIX
+from compute_space.core.env_aliases import add_bottle_env_aliases as add_bottle_env_aliases
 from compute_space.core.logging import logger
 from compute_space.core.manifest import AppManifest
-
-# Env-var naming contract.  The project was renamed OpenHost -> Cloud in a
-# Bottle; every OPENHOST_* variable is now also exposed under BOTTLE_*.  The
-# legacy names are kept indefinitely for backward compatibility.
-LEGACY_ENV_PREFIX = "OPENHOST_"
-ENV_PREFIX = "BOTTLE_"
-
-
-def add_bottle_env_aliases(env: dict[str, str]) -> dict[str, str]:
-    """Return ``env`` with a ``BOTTLE_``-prefixed twin for every ``OPENHOST_`` var.
-
-    The rename from OpenHost to Cloud in a Bottle keeps the legacy
-    ``OPENHOST_*`` names for compatibility while exposing the same values
-    under ``BOTTLE_*``.  An already-present ``BOTTLE_*`` entry is never
-    clobbered, so an explicit new-style value wins over the auto-alias.
-    """
-    aliased = dict(env)
-    for key, value in env.items():
-        if key.startswith(LEGACY_ENV_PREFIX):
-            twin = ENV_PREFIX + key[len(LEGACY_ENV_PREFIX) :]
-            aliased.setdefault(twin, value)
-    return aliased
 
 
 def rmtree_with_sudo_fallback(path: str, *, raise_on_failure: bool = False) -> None:
